libtrx/room: move common FD handling to libtrx
This moves the common floor data parsing routines, consts and macros to
libtrx.
