+++ src/libtrx/include/libtrx/game/rooms/common.h
@@ -10,3 +10,8 @@ extern bool Room_GetFlipStatus(void);
 
 int32_t Room_GetAdjoiningRooms(
     int16_t init_room_num, int16_t out_room_nums[], int32_t max_room_num_count);
+
+void Room_ParseFloorData(const int16_t *floor_data);
+void Room_PopulateSectorData(
+    SECTOR *sector, const int16_t *floor_data, uint16_t start_index,
+    uint16_t null_index);
