+++ src/tr2/game/room.h
@@ -27,11 +27,6 @@ int32_t __cdecl Room_GetHeight(
 int32_t __cdecl Room_GetCeiling(
     const SECTOR *sector, int32_t x, int32_t y, int32_t z);
 
-void Room_ParseFloorData(const int16_t *floor_data);
-void Room_PopulateSectorData(
-    SECTOR *sector, const int16_t *floor_data, uint16_t start_index,
-    uint16_t null_index);
-
 void Room_TestTriggers(const ITEM *item);
 void __cdecl Room_AlterFloorHeight(const ITEM *item, int32_t height);
 void __cdecl Room_FlipMap(void);
